+++ lib/components/LinearProgressIndicator/LinearProgressIndicator.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react';
 import * as React from 'react';
 
-import { LinearProgressIndicator } from '.';
+import { LinearProgressIndicator } from './LinearProgressIndicator';
 
 export default {
 	title: 'Components/Progress/Linear',
